pyproject.toml
[project]
name = "c2cli"
version = "0.1.0"
description = "Computer Vision Format Converter - Convert between COCO, YOLO, and Pascal VOC annotation formats"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["computer-vision", "annotation", "coco", "yolo", "pascal-voc", "object-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
c2cli = "c2cli:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
